+++ ouzel/input/ios/InputSystemIOS.hpp
@@ -44,7 +44,7 @@ namespace ouzel
             void showVirtualKeyboard();
             void hideVirtualKeyboard();
 
-            uint32_t lastDeviceId = 0;
+            uintptr_t lastDeviceId = 0;
             std::unique_ptr<KeyboardDevice> keyboardDevice;
             std::unique_ptr<TouchpadDevice> touchpadDevice;
             std::unordered_map<GCControllerPtr, std::unique_ptr<GamepadDeviceIOS>> gamepadDevices;
